Revert "adding a state and container class"
