+++ container.py
@@ -1,24 +1,18 @@
-from typing import Optional
-
-class Container:
-    def __init__(self,
-                 position: Optional[int] = None,
-                 weight: Optional[float] = None,
-                 description: Optional[str] = None) -> None:
-        
-        self.position = position
+class container:
+    def __init__(self, location, weight, description):
+        self.location = location
         self.weight = weight
         self.description = description
-
-    def get_weight(self):
-        return self.weight
-
-    def get_description(self):
-        return self.description
-    
-    def get_position(self):
-        return self.position
-
-    def print_node_description(self):
-        return f"Node(weight={self.weight}, description={self.description}, position={self.position})"
-
+        
+        if description == "NAN":
+            self.IsNAN = True
+        else:
+            self.IsNAN = False
+            
+        if description == "UNUSED":
+            self.IsUNUSED = True
+        else:
+            self.IsUNUSED = False
+            
+    def printContainer(self):
+        print(self.location, self.weight, self.description)
