+++ log.txt
@@ -2,29 +2,3 @@ November 6th 2024: 15:28 Insert Message to log.txt
 November 6th 2024: 15:28 Manifest sampleManifest.txt is opened
 November 6th 2024: 15:31 Insert Message to log.txt
 November 6th 2024: 15:31 Manifest sampleManifest.txt is opened
-November 20th 2024: 21:57 Insert Message to log.txt
-November 20th 2024: 21:57 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:38 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:39 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:40 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:41 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:41 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:42 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:42 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:45 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:46 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:46 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:46 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:46 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:46 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:47 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:47 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:47 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:47 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:47 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:48 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:49 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:49 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:49 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:49 Manifest sampleManifest.txt is opened
-November 20th 2024: 22:49 Manifest sampleManifest.txt is opened
